Decompose StrategyBuilder (1124 -> 94 lines) + extract & test its logic
StrategyBuilder.tsx was a 1124-line monolith: one component held a 250-line
metric computation buried in a useMemo plus ~760 lines of JSX across eight
unrelated visual sections. It had no tests and mixed business logic with view.

Split into src/components/strategy-builder/:
- computeCombinedMetrics.ts: the aggregation/weighting/risk math as a pure,
  typed function (CombinedMetrics), lifted verbatim -- same math, including the
  existing Math.random()-simulated series (behaviour preserved, documented).
- chartTheme.ts: shared COLORS / STRATEGY_COLORS.
- StrategySelection, PerformanceOverview, PerformanceCharts, AllocationCharts,
  HoldingsConcentration, AdvancedSections, StrategySummary, HoldingsModal:
  presentational components, each fed the computed model + theme.

StrategyBuilder.tsx is now a 94-line orchestrator: state + one memoized call to
computeCombinedMetrics + composition.

Dead code removed along the way (all previously unused): searchTerm/setSearchTerm,
selectAll, clearAll, filteredStrategies, the SectionHeader inner component, and
four never-toggled expandedSections keys (kept only diversification/trading/
leverage, which are the only sections that exist).

Tests: adds vitest (first front-end test runner here) + a `test` script, and 8
tests on computeCombinedMetrics (totals, selected-subset filtering, asset
allocation + the <3% "Others" bucket, value-weighted vol / max drawdown,
per-symbol PnL, unit-diagonal correlation matrix). All pass. `vite build`
succeeds.
