+++ src/components/strategy-builder/chartTheme.ts
@@ -1,0 +1,9 @@
+// Shared palettes for the StrategyBuilder charts and tables. Kept in one place so
+// the pie slices, cell dots and legends stay in sync across the sub-components.
+
+export const COLORS = [
+  '#f97316', '#fb923c', '#fdba74', '#fed7aa', '#ffedd5',
+  '#3b82f6', '#60a5fa', '#93c5fd', '#dbeafe'
+];
+
+export const STRATEGY_COLORS = ['#f97316', '#fb923c', '#fdba74'];
